go.codereview: bump repo to get builders going again
The dashboard has stopped accepting new builds for this sub repo revision. This causes the builders to loop contiously burning CPU and wearing out flash partitions.

The workaround, for the moment, is to bump the revision to get things going again. See http://codereview.appspot.com/6847124 for details.

 2013/02/04 20:19:17 buildSubrepos code.google.com/p/go.codereview: AddResult: putting Commit: Call error 5: SharedBuffer request too large for API.

R=adg, dsymonds
CC=golang-dev
https://codereview.appspot.com/7284048